Pineapple added for some roles

+++ src/events/messageCreate.ts
@@ -3,7 +3,8 @@ import { Message } from 'discord.js';
 export default {
   name: 'messageCreate',
   async execute (message: Message) {
-    if (message.content.includes('🍍')) {
+    const member = message.member;
+    if (message.content.includes('🍍') && !member?.roles.cache.get('1088349504094359572')  && !member?.roles.cache.get('1048590352845635595')) {
       await message.delete();
     }
   }
